+++ src/main/java/roomescape/RoomescapeApplication.java
@@ -8,5 +8,4 @@ public class RoomescapeApplication {
     public static void main(String[] args) {
         SpringApplication.run(RoomescapeApplication.class, args);
     }
-
 }
