+++ src/main/java/roomescape/domain/Reservation.java
@@ -1,13 +1,13 @@
-package roomescape;
+package roomescape.domain;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.LocalTime;
 
 public class Reservation {
-     private final long id;
-     private String name;
-     private LocalDateTime dateTime;
+    private final long id;
+    private String name;
+    private LocalDateTime dateTime;
 
     public Reservation(long id, String name, LocalDateTime dateTime) {
         this.id = id;
