+++ src/main/java/roomescape/dto/ReservationResponse.java
@@ -1,8 +1,9 @@
-package roomescape;
+package roomescape.dto;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import java.time.LocalDate;
 import java.time.LocalTime;
+import roomescape.domain.Reservation;
 
 public record ReservationResponse(long id, String name, LocalDate date,
                                   @JsonFormat(pattern = "HH:mm") LocalTime time) {
